// src/app/data/model/dataTestPredefinidos.ts
export interface DataTestPredefinido {
    cdicurso: number;
    id_curso: string;
    nombre_curso: string;
    cdipermiso: number;
    nombre_permiso: string;
    cdicategoria: number;
    nombre_categoria: string;
    cditest: number;
    nombre_test: string;
    descripcion_test: string;
    traducir: number;
    idioma: string;
    ayuda: number;
    autocorreccion: number;
    propia: number;
}